Rename frontend CityList module to City
