+++ frontend/src/components/City/EditListItem.js
@@ -9,7 +9,7 @@ import { Formik } from 'formik';
 import { withStyles } from '@material-ui/core/styles';
 import { Mutation } from 'react-apollo';
 import gql from 'graphql-tag';
-import { allCitiesQuery } from '.';
+import { allCitiesQuery } from './List';
 
 const styles = theme => ({
   textField: {
