+++ frontend/src/pages/admin/cidades.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import AppFrame from '../../components/AppFrame';
 import AppContent from '../../components/AppContent';
-import CityList from '../../components/CityList';
+import CityList from '../../components/City/List';
 
 const PageCities = () => (
   <AppFrame>
